% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{calculate_codon_density}
\alias{calculate_codon_density}
\title{Calculate codon density}
\usage{
calculate_codon_density(
  bam_dat,
  transcript_length,
  exclude_codons5 = 10,
  exclude_codons3 = 10,
  normalize = F,
  which_column = "count"
)
}
\arguments{
\item{bam_dat}{data frame; output from `load_bam`}

\item{transcript_length}{integer; length of transcript CDS in codons}

\item{exclude_codons5}{integer; number of codons to exclude from 5' end of transcript}

\item{exclude_codons3}{integer; number of codons to exclude from 3' end of transcript}

\item{normalize}{logical; whether to normalize codon counts by the average RPF count across the transcript}

\item{which_column}{character; column in `bam_dat` corresponding to RPF count}
}
\value{
A numeric vector of RPF counts per codon position
}
\description{
This function returns a vector of RPF counts per codon for an individual transcrips.
}
