% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{calculate_transcript_density}
\alias{calculate_transcript_density}
\title{Calculate per-transcript RPF density}
\usage{
calculate_transcript_density(
  bam_dat,
  transcript_length_fname,
  statistic = mean,
  exclude_codons5 = 10,
  exclude_codons3 = 10
)
}
\arguments{
\item{bam_dat}{data frame; output from `load_bam`}

\item{transcript_length_fname}{character; file path to transcriptome lengths file}

\item{statistic}{character; one of `mean` or `median`}

\item{exclude_codons5}{integer; number of codons to exclude from 5' end of transcript}

\item{exclude_codons3}{integer; number of codons to exclude from 3' end of transcript}
}
\value{
A numeric vector of RPF density per transcript
}
\description{
This function computes the mean or median footprint density per codon across
all transcripts.
}
