% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{choose_subsets}
\alias{choose_subsets}
\title{Choose RPF lengths and frames to model}
\usage{
choose_subsets(d5_d3, min_prop = 0.9)
}
\arguments{
\item{d5_d3}{list; output from `count_d5_d3`}

\item{min_prop}{numeric; minimum proportion of RPF counts to be modeled}
}
\value{
A data frame of 5' and 3' digest lengths to be used for modeling
RPF counts.
}
\description{
This function chooses the most populous RPF 5' and 3' length combinations
to model some minimum proportion of RPF counts.
}
