% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choros.R
\name{choros}
\alias{choros}
\title{Pipeline for RPF bias correction}
\usage{
choros(
  bam_fname,
  transcript_fa_fname,
  transcript_length_fname,
  offsets_fname,
  expt_dir = ".",
  save_intermediate = T,
  extra_args = NULL
)
}
\arguments{
\item{bam_fname}{character; file path to .bam alignment file}

\item{transcript_fa_fname}{character; file path to transcriptome .fasta file}

\item{transcript_length_fname}{character; file path to transcriptome lengths file}

\item{offsets_fname}{character; file path to offset / A site assignment rules .txt file}

\item{expt_dir}{character; file path to directory to save output/intermediate files}

\item{save_intermediate}{logical; whether to save intermediate files (if FALSE, 
only save alignment file with corrected weights)}

\item{extra_args}{named list; additional parameters to pass onto internal functions}
}
\value{
A list containing: data frame of all RPF alignments with bias-
corrected counts; data frame used for training the regression model; 
regression coefficients from the model fit; and a list of generated plots
(diagnostic plot of RPF lengths and frames; bias evaluation per-codon and 
per-nucleotide position before and after bias correction)
}
\description{
This is a wrapper function to perform all the steps in the `choros` pipeline: 
pre-processing of the RPF alignments, deciding which RPF lengths and frames
to model, deciding which transcripts to use for training, regression 
modeling, correction of RPF counts, and evaluation of position-based
biases before/after correction.
}
