% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compute_rpf_gc}
\alias{compute_rpf_gc}
\title{Compute GC-content for a ribosome-protected fragment}
\usage{
compute_rpf_gc(
  dat,
  omit = "APE",
  transcript_fa_fname,
  transcript_lengths_fname,
  read_type = "monosome"
)
}
\arguments{
\item{dat}{data frame containing columns `transcript`, `cod_idx`, `d5`, and `d3`}

\item{omit}{character; one of `A`, `AP`, or `APE` corresponding to codons to omit}

\item{transcript_fa_fname}{character; file path to transcriptome fasta file}

\item{transcript_lengths_fname}{character; file path to transcript lengths file}

\item{read_type}{character; one of `monosome` or `disome`}
}
\value{
A numeric vector corresponding to rows of `dat`
}
\description{
This function computes the GC-content of a ribosome-protected fragment, 
omitting the nucleotides in the A-, P-, and/or E-site codon positions.
}
