% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_bias.R
\name{correct_bias}
\alias{correct_bias}
\alias{correct_interxn}
\alias{correct_marginal}
\alias{correct_fixed}
\alias{correct_gc}
\title{Correct for sequence biases in RPF counts}
\usage{
correct_bias(
  dat,
  fit_coefs,
  f5_method = NULL,
  f3_method = NULL,
  correct_gc = T,
  which_column = "count"
)

correct_interxn(dat, which_column, which_region, fit_coefs)

correct_marginal(dat, which_column, which_region, fit_coefs)

correct_fixed(dat, which_column, which_region, bias_params)

correct_gc(dat, which_column, fit_coefs)
}
\arguments{
\item{dat}{data frame containing regression predictors}

\item{fit_coefs}{data frame; output from parse_coefs()}

\item{f5_method}{character; name of function to apply for 5' correction}

\item{f3_method}{character; name of function to apply for 3' correction}

\item{correct_gc}{logical; whether to correct for RPF %GC}

\item{which_column}{character; name of column in `dat` containing uncorrected counts}

\item{which_region}{character; one of "f5" or "f3" corresponding to RPF region}

\item{bias_params}{named numeric vector; recovery probabilities from simulation}
}
\value{
A numeric vector corresponding to `dat` containing bias-corrected RPF counts
}
\description{
`correct_bias` is a wrapper function that computes and applies bias correction factors 
from extracted regression coefficients.

`correct_interxn` corrects RPF counts using bias correction factors generated from
user-provided regression coefficients; this function will use regression
coefficients corresponding to interaction terms.

`correct_marginal` corrects RPF counts using bias correction factors generated from
user-provided regression coefficients; this function will ignore regression
coefficients corresponding to interaction terms.

`correct_fixed` corrects RPF counts using user-specified values; this function
is intended to be used with values from a simulation experiment.

`correct_gc` calculates a scaling factor for RPF GC content using 
user-provided regression coefficients and provide \\%GC-corrected RPF counts.
}
\section{Functions}{
\itemize{
\item \code{correct_interxn()}: Use regression coefficients as scaling factors for bias correction

\item \code{correct_marginal()}: Use regression coefficients as scaling factors for bias correction

\item \code{correct_fixed()}: Use simulation parameters as bias correction factors

\item \code{correct_gc()}: Correct for RPF %GC

}}
