% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{count_d5_d3}
\alias{count_d5_d3}
\title{Count RPFs by 5' and 3' digest lengths}
\usage{
count_d5_d3(bam_dat, plot_title = "")
}
\arguments{
\item{bam_dat}{data frame; output from `load_bam`}

\item{plot_title}{character; title for heatmap plot}
}
\value{
A list containing a data frame of RPF counts by 5' and 3' digest
lengths, and ggplot heatmap of RPF counts
}
\description{
Count number of RPFs per combination of 5' and 3' digest lengths
}
