% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{count_footprints}
\alias{count_footprints}
\title{Count RPFs by position, frame, and length}
\usage{
count_footprints(
  bam_dat,
  regression_data,
  which_column = "count",
  features = c("transcript", "cod_idx", "d5", "d3"),
  integer_counts = T
)
}
\arguments{
\item{bam_dat}{data frame; output from `load_bam`}

\item{regression_data}{data frame; output from `init_data`}

\item{which_column}{character; name of column containing counts in `bam_dat`}

\item{features}{character vector; terms to match between `regression_data` and `bam_dat`}

\item{integer_counts}{logical; whether to round counts to integers}
}
\value{
A numeric vector of RPF counts corresponding to rows in `regression_data`
}
\description{
This function aggregates RPF counts for a given transcript, codon position,
and 5' and 3' digest lengths (which correspond to an RPF length and frame).
}
