% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{count_nonzero_codons}
\alias{count_nonzero_codons}
\title{Count nonzero codon positions per transcript}
\usage{
count_nonzero_codons(bam_dat)
}
\arguments{
\item{bam_dat}{data frame; output from `load_bam`}
}
\value{
A named numeric vector
}
\description{
This function retruns the number of codon positions within a transcript with
non-zero RPF counts.
}
