% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_bias.R
\name{evaluate_bias}
\alias{evaluate_bias}
\title{Evaluate positional contributions to RPF counts}
\usage{
evaluate_bias(
  dat,
  which_column = "count",
  transcripts_fa_fname,
  transcripts_length_fname,
  trunc5 = 20,
  trunc3 = 20,
  num_f5_codons = 6,
  num_f3_codons = 6,
  type = "codon",
  metric = "corr"
)
}
\arguments{
\item{dat}{data frame containing regression predictors and column of RPF counts}

\item{which_column}{character; name of column containing counts}

\item{transcripts_fa_fname}{character; file path to transcriptome .fasta file}

\item{transcripts_length_fname}{character; file path to transcriptome lengths file}

\item{trunc5}{integer; number of codons (starting at start codon) to omit in codon correlation regression}

\item{trunc3}{integer; number of codons (starting at stop codon) to omit in codon correlation regression}

\item{num_f5_codons}{integer; number of codons 5' of A site to include in regression}

\item{num_f3_codons}{integer; number of codons 3' of A site to include in regression}

\item{type}{character; one of `codon` or `nt`}

\item{metric}{character; one of `corr` or `norm`}
}
\value{
A named numeric vector of positional contributions to RPF count. 
Values will correspond to either the decrease in correlation between predicted 
and true RPF counts in the leave-one-out model (as computed in the `iXnos` model),
or the norm of regression coefficients at that position.
}
\description{
This function will perform a series of leave-one-out regressions to evaluate
the contribution of individual positions to overall RPF count.
}
