% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_bias_seq}
\alias{get_bias_seq}
\title{Retrieve nucleotides at RPF ends}
\usage{
get_bias_seq(
  dat,
  transcript_seq,
  bias_region,
  bias_length = 3,
  read_type = "monosome"
)
}
\arguments{
\item{dat}{data frame containing columns `transcript`, `d5`, `d3`, `utr5_length`, and codon indices.}

\item{transcript_seq}{character; output from `load_fasta`}

\item{bias_region}{character; one of `f5` or `f3` corresponding to 5' or 3' bias sequence}

\item{bias_length}{integer; number of nucleotides to include in bias sequence}

\item{read_type}{character; one of `monosome` or `disome`}
}
\value{
A character vector corresponding to rows of `dat`
}
\description{
A function to pull the nucleotides at the ends of a ribosome-protected 
fragment, defined by ribosome position and size.
}
\details{
If `read_type` is `monosome`, then A site codon indices should be named 
`cod_idx` in `dat`. If `read_type` is `disome`, then A site codon indices 
should be named `cod_idx_lagging` and `cod_idx_leading` in `dat`.
}
