% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_codons}
\alias{get_codons}
\title{Retrieve codons in A-, P-, and E-site positions}
\usage{
get_codons(transcript_name, cod_idx, utr5_length, transcript_seq)
}
\arguments{
\item{transcript_name}{character; one of `names(transcript_seq)`}

\item{cod_idx}{integer; codon index for A-site codon (1-based)}

\item{utr5_length}{integer; length of 5' UTR (in nucleotides)}

\item{transcript_seq}{named character vector; transcript (+ 5' and 3' UTR regions) sequences}
}
\value{
A named character vector
}
\description{
A function to pull codons in the A-, P-, and E-site positions of a ribosome
at a position within a transcript.
}
