% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{init_data}
\alias{init_data}
\title{Generate data frame to fit negative binomial regression model}
\usage{
init_data(
  transcript_fa_fname,
  transcript_length_fname,
  digest5_lengths = 15:18,
  digest3_lengths = 9:11,
  d5_d3_subsets = NULL,
  f5_length = 3,
  f3_length = 3,
  num_cores = NULL,
  which_transcripts = NULL,
  exclude_codons5 = 10,
  exclude_codons3 = 10,
  compute_gc = T,
  gc_omit = "APE"
)
}
\arguments{
\item{transcript_fa_fname}{character; file path to transcriptome .fasta file}

\item{transcript_length_fname}{character; file path to transcriptome lengths file}

\item{digest5_lengths}{integer vector; legal 5' digest lengths}

\item{digest3_lengths}{integer vector; legal 3' digest lengths}

\item{d5_d3_subsets}{data frame; columns of `d5` and `d3` to initiate data over}

\item{f5_length}{integer; length of 5' bias sequence}

\item{f3_length}{integer; length of 3' bias sequence}

\item{num_cores}{integer; number of cores to use for parallelization}

\item{which_transcripts}{character vector; transcripts selected for regression}

\item{exclude_codons5}{integer; number of codons to exclude from 5' end of transcript}

\item{exclude_codons3}{integer; number of codons to exclude from 3' end of transcript}

\item{compute_gc}{logical; whether to return RPF GC-content}

\item{gc_omit}{character; one of `A`, `AP`, or `APE` codons to omit for GC-content calculation}
}
\value{
data frame of data to use for downstream regression modeling
}
\description{
This function enumerates data points to train a regression model of RPF counts,
given a list of transcripts for training
}
