% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{load_bam}
\alias{load_bam}
\title{Pre-process RPF alignment file}
\usage{
load_bam(
  bam_fname,
  transcript_fa_fname,
  transcript_length_fname,
  offsets_fname = NULL,
  num_cores = NULL,
  gc_omit = "APE",
  read_type = "monosome",
  f5_length = 3,
  f3_length = 3,
  offsets_5prime_fname = NULL,
  offsets_3prime_fname = NULL
)
}
\arguments{
\item{bam_fname}{character; file path to .bam alignment file}

\item{transcript_fa_fname}{character; file path to transcriptome .fasta file}

\item{transcript_length_fname}{character; file path to transcriptome lengths file}

\item{offsets_fname}{character; file path to A-site assignment rules file}

\item{num_cores}{integer; number of cores to use for parallelization}

\item{read_type}{character; one of `monosome` or `disome`}

\item{f5_length}{integer; length of sequence between RPF 5' end and A site}

\item{f3_length}{integer; length of sequence between A site and RPF 3' end}
}
\value{
A data frame of aggregated RPF counts and annotations
}
\description{
This function takes a .bam alignment file where each row corresponds to
an individual RPF read; generates annotations for RPF codon index, 5'
digest length, 3' digest length, A-/P-/E-site codons, GC content, and bias
sequences; filters out invalid RPFs; and aggregates identical RPFs.

If `read_type` is `disome`, user must also specify `offsets_5prime_fname`
and `offsets_3prime_fname`.
}
