% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{match_rows}
\alias{match_rows}
\title{Get indices for corresponding rows in two data frames}
\usage{
match_rows(x, y, x_col = NULL, y_col = NULL)
}
\arguments{
\item{x}{data frame containing rows to find in `y`}

\item{y}{data frame}

\item{x_col}{character vector of columns to be used; `NULL` to use all columns}

\item{y_col}{character vector of columns corresponding to `x_col`; `NULL` to use `x_col`}
}
\value{
Numeric vector corresponding to rows in `x`
}
\description{
Return row indices of (first) matches of `x` in `y`
}
