% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parse_coefs}
\alias{parse_coefs}
\title{Annotate regression coefficients}
\usage{
parse_coefs(nb_fit)
}
\arguments{
\item{nb_fit}{`negbin` object from `glm.nb`}
}
\value{
A data frame containing regression coefficients and corresponding annotations.
}
\description{
Extract regression coefficients from a negative binomial object and annotate
coefficient types for downstream analyses.
}
