% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_bias}
\alias{plot_bias}
\title{Plot RPF biases}
\usage{
plot_bias(
  model_metric,
  plot_title = "",
  plot_subtitle = "",
  type = "codon",
  metric = "corr",
  fill_colors = NULL
)
}
\arguments{
\item{model_metric}{numeric vector; output from `evaluate_bias`}

\item{plot_title}{character; title for plot}

\item{plot_subtitle}{character; subtitle for plot}

\item{type}{character; one of `codon` or `nt` corresponding to `model_metric`}

\item{metric}{character; one of `corr` or `norm`}

\item{fill_colors}{named character vector; colors corresponding to positions}
}
\value{
A ggplot object
}
\description{
This function generates `iXnos`-style plot of codon or nucleotide positional 
contributions to RPF counts.
}
