% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Visualize regression coefficients}
\usage{
plot_coefs(
  model_coefs,
  plot_type = "violin",
  volcano_term = "A",
  conf_int = 0.95,
  p_adj_method = "fdr",
  p_adj_group = c("all", "subset_only")
)
}
\arguments{
\item{model_coefs}{data frame; output from `parse_coefs`}

\item{plot_type}{character; one of `violin` or `volcano`}

\item{volcano_term}{character; coefficient group to plot}

\item{conf_int}{numeric; threshold for statistical significance}

\item{p_adj_method}{character; `method` argument for `p.adjust` for multiple hypothesis correction}

\item{p_adj_group}{character; one of `all` or `subset_only`}
}
\description{
This function generates either a violin plot of regression coefficient values
or a volcano plot of regression coefficient values by p-value.
}
