% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Generate diagnostic plots}
\usage{
plot_diagnostic(
  bam_fname,
  transcript_length_fname,
  plot_title = "",
  start_min = -20,
  start_max = 5,
  stop_min = -5,
  stop_max = 20,
  length_min = 15,
  length_max = 35
)
}
\arguments{
\item{bam_fname}{character; file path to .bam alignment file}

\item{transcript_length_fname}{character; file path to transcriptome lengths file}

\item{plot_title}{character; title for diagnostic plot}

\item{start_min}{integer; 5'-most position to include in plot, relative to start codon}

\item{start_max}{integer; 3'-most position to include in plot, relative to start codon}

\item{stop_min}{integer; 5'-most position to include in plot, relative to stop codon}

\item{stop_max}{integer; 3'-most position to include in plot, relative to stop codon}

\item{length_min}{integer; minimum RPF length}

\item{length_max}{integer; maximum RPF length}
}
\value{
A patchwork object of diagnostic plots
}
\description{
This function generates diagnostic plots: RPF density at the start
codon; RPF density at the stop codon; histogram of RPF lengths; heatmap of 
RPF counts by RPF length and frame

Input `.bam` file must contain `tag.ZW` of posterior weights computed from
`RSEM`
}
