% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_fasta}
\alias{write_fasta}
\title{Write transcriptome sequence to disk}
\usage{
write_fasta(transcript_seq, fa_fname)
}
\arguments{
\item{transcript_seq}{named character vector; file path to output .fasta file}
}
\description{
Write a character vector of transcript sequences to a .fasta file. Names of 
`transcript_seq` will become FASTA headers.
}
